refined script and updated workflow to use the script
[skip ci]

Jira ticket: CAMS-421
