+++ .github/workflows/reusable-e2e.yml
@@ -65,6 +65,12 @@ jobs:
           -p ${{ env.priority }} \
           --is-cicd
 
+      - name: Install Node dependencies
+        working-directory: test/e2e
+        run: |
+          npm ci
+          npx playwright install chromium msedge
+
       - name: Seed e2e database ##If E2E CosmosDB exists and has been seeded, we do not want to seed again
         if: ${{ inputs.e2eCosmosDbExists == 'false' }}
         run: |
@@ -74,14 +80,19 @@ jobs:
             e2eDatabaseName="${e2eDatabaseName}-${{ inputs.branchHashId }}"
           fi
 
-          echo "Seeding E2E CosmosDB..."
-          curl -v -d '{"txIdOverride": '0'}' -H "Content-Type: application/json" https://${{ inputs.apiFunctionName }}.azurewebsites.us/api/orders-sync?x-ms-routing-name=${{ inputs.slotName }}
+          export MSSQL_HOST=${{ secrets.MSSQL_HOST }}
+          export MSSQL_DATABASE_DXTR=${{ secrets.MSSQL_DATABASE_DXTR }}
+          export MSSQL_USER=${{ secrets.MSSQL_USER }}
+          export MSSQL_PASS=${{ secrets.MSSQL_PASS }}
+          export MSSQL_ENCRYPT=true
+          export MSSQL_TRUST_UNSIGNED_CERT=true
+          export MONGO_CONNECTION_STRING="${{ secrets.MONGO_CONNECTION_STRING }}"
+          export COSMOS_DATABASE_NAME="${e2eDatabaseName}"
+          export DATABASE_MOCK=false
 
-      - name: Install Node dependencies
-        working-directory: test/e2e
-        run: |
-          npm ci
-          npx playwright install chromium msedge
+          echo "Seeding E2E CosmosDB..."
+
+          cd ./test/e2e && npm run load-data
 
       - name: Run E2E Tests
         id: e2eTests
